Rename open_graph.py to opengraph.py
